repos/devin1/modules/cloud_integration_services.py
"""
cloud_integration_services.py

This module handles the management of various cloud services, including
deployment, monitoring, and resource scaling. It supports multiple cloud platforms
such as AWS, Google Cloud Platform, Azure, and more.
"""

import boto3
from google.cloud import storage
from azure.identity import DefaultAzureCredential
from azure.mgmt.resource import ResourceManagementClient
import logging

logging.basicConfig(level=logging.INFO)

class CloudServiceManager:
    """
    A class to manage cloud services across multiple platforms.
    """

    def __init__(self):
        self.aws_client = None
        self.gcp_client = None
        self.azure_client = None

    # AWS Management Methods
    def configure_aws(self, access_key: str, secret_key: str, region: str):
        """
        Configure AWS services.
        """
        self.aws_client = boto3.client(
            'ec2',
            aws_access_key_id=access_key,
            aws_secret_access_key=secret_key,
            region_name=region,
        )
        logging.info("AWS client configured.")

    def list_aws_instances(self):
        """
        List all EC2 instances in AWS.
        """
        if not self.aws_client:
            raise ValueError("AWS client is not configured.")
        instances = self.aws_client.describe_instances()
        return instances

    # GCP Management Methods
    def configure_gcp(self, service_account_file: str):
        """
        Configure GCP services.
        """
        self.gcp_client = storage.Client.from_service_account_json(service_account_file)
        logging.info("GCP client configured.")

    def list_gcp_buckets(self):
        """
        List all storage buckets in GCP.
        """
        if not self.gcp_client:
            raise ValueError("GCP client is not configured.")
        buckets = [bucket.name for bucket in self.gcp_client.list_buckets()]
        return buckets

    # Azure Management Methods
    def configure_azure(self, subscription_id: str):
        """
        Configure Azure services.
        """
        credentials = DefaultAzureCredential()
        self.azure_client = ResourceManagementClient(credentials, subscription_id)
        logging.info("Azure client configured.")

    def list_azure_resources(self):
        """
        List all resources in Azure.
        """
        if not self.azure_client:
            raise ValueError("Azure client is not configured.")
        resources = self.azure_client.resources.list()
        return [res.name for res in resources]

    # Common Methods
    def deploy_service(self, platform: str, configuration: dict):
        """
        Deploy a service to the specified cloud platform.
        """
        if platform == "AWS":
            return self.deploy_aws_service(configuration)
        elif platform == "GCP":
            return self.deploy_gcp_service(configuration)
        elif platform == "Azure":
            return self.deploy_azure_service(configuration)
        else:
            raise ValueError(f"Unsupported platform: {platform}")

    def deploy_aws_service(self, configuration: dict):
        """
        Deploy a service in AWS.
        """
        # Example placeholder for actual deployment logic
        logging.info("Deploying AWS service with configuration:")
        logging.info(configuration)
        return "AWS service deployed."

    def deploy_gcp_service(self, configuration: dict):
        """
        Deploy a service in GCP.
        """
        # Example placeholder for actual deployment logic
        logging.info("Deploying GCP service with configuration:")
        logging.info(configuration)
        return "GCP service deployed."

    def deploy_azure_service(self, configuration: dict):
        """
        Deploy a service in Azure.
        """
        # Example placeholder for actual deployment logic
        logging.info("Deploying Azure service with configuration:")
        logging.info(configuration)
        return "Azure service deployed."

if __name__ == "__main__":
    cloud_manager = CloudServiceManager()

    # Example usage
    # AWS
    cloud_manager.configure_aws("your_access_key", "your_secret_key", "us-west-1")
    print(cloud_manager.list_aws_instances())

    # GCP
    cloud_manager.configure_gcp("path_to_service_account.json")
    print(cloud_manager.list_gcp_buckets())

    # Azure
    cloud_manager.configure_azure("your_subscription_id")
    print(cloud_manager.list_azure_resources())
